many: move proxy to cmd/snapd
