+++ cmd/snapd/snap-userdb-proxy/testhelpers_test.go
